% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lst.R
\name{calculate_lst}
\alias{calculate_lst}
\title{Calculate LST}
\usage{
calculate_lst(
  B4,
  B5,
  B10,
  B11,
  pv_coeff = 0.004,
  lse_coeff = 0.986,
  unit = "K"
)
}
\arguments{
\item{B4}{Raster object representing the B4 (red) band.}

\item{B5}{Raster object representing the B5 (NIR) band.}

\item{B10}{Raster object representing the B10 (thermal) band.}

\item{B11}{Raster object representing the B11 (thermal) band.}

\item{pv_coeff}{Coefficient for PV calculation. Default is 0.004.}

\item{lse_coeff}{Coefficient for LSE calculation. Default is 0.986.}

\item{unit}{Character string indicating the output unit ("K" for Kelvin, "C" for Celsius). Default is "K".}
}
\value{
A `raster` object representing the LST.
}
\description{
Calculates Land Surface Temperature (LST) from Landsat 8 bands B10 and B11.
}
\examples{
\dontrun{
B4_path <- system.file("extdata", "B4.tif", package = "DroughtIndices")
B5_path <- system.file("extdata", "B5.tif", package = "DroughtIndices")
B10_path <- system.file("extdata", "B10.tif", package = "DroughtIndices")
B11_path <- system.file("extdata", "B11.tif", package = "DroughtIndices")
B4 <- raster(B4_path)
B5 <- raster(B5_path)
B10 <- raster(B10_path)
B11 <- raster(B11_path)
lst <- calculate_lst(B4, B5, B10, B11)
print(lst)
}
}
