% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ndvi.R
\name{calculate_ndvi}
\alias{calculate_ndvi}
\title{Calculate NDVI}
\usage{
calculate_ndvi(B4, B5)
}
\arguments{
\item{B4}{Raster object representing the B4 (red) band.}

\item{B5}{Raster object representing the B5 (NIR) band.}
}
\value{
A `raster` object representing the NDVI.
}
\description{
Calculates NDVI from the red (B4) and NIR (B5) bands.
}
\examples{
\dontrun{
B4_path <- system.file("extdata", "B4.tif", package = "DroughtIndices")
B5_path <- system.file("extdata", "B5.tif", package = "DroughtIndices")
B4 <- raster(B4_path)
B5 <- raster(B5_path)
ndvi <- calculate_ndvi(B4, B5)
print(ndvi)
}
}
