% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_precipitation_index.R
\name{calculate_precipitation_index}
\alias{calculate_precipitation_index}
\title{Calculate Precipitation Index}
\usage{
calculate_precipitation_index(precipitation_raster)
}
\arguments{
\item{precipitation_raster}{Raster object representing the precipitation data.}
}
\value{
A `raster` object representing the PCI.
}
\description{
Reads and processes precipitation data to calculate PCI.
}
\examples{
\dontrun{
precipitation_path <- system.file("extdata", "precipitation.tif", package = "DroughtIndices")
precipitation <- raster(precipitation_path)
pci <- calculate_precipitation_index(precipitation)
print(pci)
}
}
