% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sdci.R
\name{calculate_sdci}
\alias{calculate_sdci}
\title{Calculate SDCI}
\usage{
calculate_sdci(vci, tci, pci)
}
\arguments{
\item{vci}{Raster object representing the Vegetation Condition Index (VCI).}

\item{tci}{Raster object representing the Temperature Condition Index (TCI).}

\item{pci}{Raster object representing the Precipitation Condition Index (PCI).}
}
\value{
A raster object representing the SDCI.
}
\description{
Calculates the Standardized Drought Condition Index (SDCI).
}
\examples{
\dontrun{
vci <- raster(system.file("extdata", "VCI.tif", package = "DroughtIndices"))
tci <- raster(system.file("extdata", "TCI.tif", package = "DroughtIndices"))
pci <- raster(system.file("extdata", "PCI.tif", package = "DroughtIndices"))
sdci <- calculate_sdci(vci, tci, pci)
print(sdci)
}
}
