% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tci.R
\name{calculate_tci}
\alias{calculate_tci}
\title{Calculate TCI}
\usage{
calculate_tci(lst_raster)
}
\arguments{
\item{lst_raster}{Raster object representing the LST.}
}
\value{
A `raster` object representing the TCI.
}
\description{
Calculates the Temperature Condition Index (TCI) from LST.
}
\examples{
\dontrun{
B4_path <- system.file("extdata", "B4.tif", package = "DroughtIndices")
B5_path <- system.file("extdata", "B5.tif", package = "DroughtIndices")
B10_path <- system.file("extdata", "B10.tif", package = "DroughtIndices")
B11_path <- system.file("extdata", "B11.tif", package = "DroughtIndices")
B4 <- raster(B4_path)
B5 <- raster(B5_path)
B10 <- raster(B10_path)
B11 <- raster(B11_path)
lst <- calculate_lst(B4, B5, B10, B11)
tci <- calculate_tci(lst)
print(tci)
}
}
