% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_vci.R
\name{calculate_vci}
\alias{calculate_vci}
\title{Calculate VCI}
\usage{
calculate_vci(ndvi_raster)
}
\arguments{
\item{ndvi_raster}{Raster object representing the NDVI.}
}
\value{
A `raster` object representing the VCI.
}
\description{
Calculates the Vegetation Condition Index (VCI) from NDVI.
}
\examples{
\dontrun{
B4_path <- system.file("extdata", "B4.tif", package = "DroughtIndices")
B5_path <- system.file("extdata", "B5.tif", package = "DroughtIndices")
B4 <- raster(B4_path)
B5 <- raster(B5_path)
ndvi <- calculate_ndvi(B4, B5)
vci <- calculate_vci(ndvi)
print(vci)
}
}
