% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input_rasters.R
\name{read_input_rasters}
\alias{read_input_rasters}
\title{Read Input Rasters}
\usage{
read_input_rasters(B4_path, B5_path, B10_path, B11_path, precipitation_path)
}
\arguments{
\item{B4_path}{Path to the B4 (red) band raster file.}

\item{B5_path}{Path to the B5 (NIR) band raster file.}

\item{B10_path}{Path to the B10 (thermal) band raster file.}

\item{B11_path}{Path to the B11 (thermal) band raster file.}

\item{precipitation_path}{Path to the precipitation raster file.}
}
\value{
A list of `raster` objects: B4, B5, B10, B11, and precipitation.
}
\description{
Reads specified raster bands and precipitation data, and returns them as a list of `raster` objects.
}
\examples{
\dontrun{
B4_path <- system.file("extdata", "B4.tif", package = "DroughtIndices")
B5_path <- system.file("extdata", "B5.tif", package = "DroughtIndices")
B10_path <- system.file("extdata", "B10.tif", package = "DroughtIndices")
B11_path <- system.file("extdata", "B11.tif", package = "DroughtIndices")
precipitation_path <- system.file("extdata", "precipitation.tif", package = "DroughtIndices")
rasters <- read_input_rasters(B4_path, B5_path, B10_path, B11_path, precipitation_path)
print(rasters)
}
}
