% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_raster.R
\name{visualize_raster}
\alias{visualize_raster}
\title{Visualize Raster}
\usage{
visualize_raster(
  raster_layer,
  main_title = NULL,
  xlab = "Longitude",
  ylab = "Latitude",
  color_palette = c("green", "yellow", "red")
)
}
\arguments{
\item{raster_layer}{Raster object to visualize.}

\item{main_title}{Character string representing the title of the plot. Default is NULL, which generates a title based on the raster layer name.}

\item{xlab}{Character string representing the label of the x-axis. Default is "Longitude".}

\item{ylab}{Character string representing the label of the y-axis. Default is "Latitude".}

\item{color_palette}{Character vector representing the color palette. Default is c("green", "yellow", "red").}
}
\value{
A plot of the raster layer.
}
\description{
Visualize a raster layer with a specified color gradient.
}
\examples{
\dontrun{
visualize_raster(precipitation)
visualize_raster(vci)
visualize_raster(ndvi)
visualize_raster(lst)
visualize_raster(sdci)
}
}
